+++ app/src/main/java/ru/androidschool/intensiv/data/repository/FavoriteMovieRepositoryLocal.kt
@@ -7,9 +7,12 @@ import ru.androidschool.intensiv.data.dbo.MovieAndGenreAndActorAndProductionComp
 import ru.androidschool.intensiv.domain.repository.MovieRepository
 import ru.androidschool.intensiv.utils.ViewFeature
 
-object MovieRepositoryImpl : MovieRepository {
-    override fun getMovies(context: Context, viewFeature: ViewFeature): Single<List<MovieAndGenreAndActorAndProductionCompany>> {
+object FavoriteMovieRepositoryLocal : MovieRepository {
+    override fun getMovies(
+        context: Context,
+        viewFeature: ViewFeature
+    ): Single<List<MovieAndGenreAndActorAndProductionCompany>> {
         val movieDao = MovieDatabase.get(context).movieDao()
-        return movieDao.getMovies(viewFeature)
+        return movieDao.getMovies(ViewFeature.FAVORITE)
     }
 }
