+++ app/src/main/java/ru/androidschool/intensiv/data/repository/MovieRepositoryLocal.kt
@@ -7,12 +7,9 @@ import ru.androidschool.intensiv.data.dbo.MovieAndGenreAndActorAndProductionComp
 import ru.androidschool.intensiv.domain.repository.MovieRepository
 import ru.androidschool.intensiv.utils.ViewFeature
 
-object FavoriteMovieRepository : MovieRepository {
-    override fun getMovies(
-        context: Context,
-        viewFeature: ViewFeature
-    ): Single<List<MovieAndGenreAndActorAndProductionCompany>> {
+object MovieRepositoryLocal : MovieRepository {
+    override fun getMovies(context: Context, viewFeature: ViewFeature): Single<List<MovieAndGenreAndActorAndProductionCompany>> {
         val movieDao = MovieDatabase.get(context).movieDao()
-        return movieDao.getMovies(ViewFeature.FAVORITE)
+        return movieDao.getMovies(viewFeature)
     }
 }
