+++ app/src/main/java/ru/androidschool/intensiv/presentation/watchlist/WatchlistFragment.kt
@@ -6,14 +6,13 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
-import androidx.navigation.navOptions
 import androidx.recyclerview.widget.GridLayoutManager
 import com.xwray.groupie.GroupAdapter
 import com.xwray.groupie.kotlinandroidextensions.GroupieViewHolder
 import io.reactivex.disposables.CompositeDisposable
 import kotlinx.android.synthetic.main.fragment_watchlist.*
 import ru.androidschool.intensiv.R
-import ru.androidschool.intensiv.data.repository.FavoriteMovieRepository
+import ru.androidschool.intensiv.data.repository.FavoriteMovieRepositoryLocal
 
 class WatchlistFragment : Fragment(R.layout.fragment_watchlist) {
 
@@ -37,7 +36,7 @@ class WatchlistFragment : Fragment(R.layout.fragment_watchlist) {
         // 1. Создаем фабрику
         val appContext = context?.applicationContext ?: throw IllegalStateException("App context not found")
 
-        val watchlistViewModelFactory = WatchlistViewModelFactory(findNavController(), appContext, FavoriteMovieRepository)
+        val watchlistViewModelFactory = WatchlistViewModelFactory(findNavController(), appContext, FavoriteMovieRepositoryLocal)
 
         // 2. Создаем ViewModel
         watchlistViewModel = ViewModelProvider(this, watchlistViewModelFactory).get(WatchListViewModel::class.java)
